+++ src/environments/definitions/base/morpho-vaults.ts
@@ -33,13 +33,13 @@ export const vaults = createVaultConfig({
       vaultToken: "meUSDC",
       campaignId:
         "0x6738320fdf80785ff7a1d45ed93a6ffa07068ce9ec4170c1887d09f32fba7b57",
+      version: 2,
     },
-    meUSDCv2: {
+    meUSDCv1: {
       underlyingToken: "USDC",
-      vaultToken: "meUSDCv2",
+      vaultToken: "meUSDCv1",
       campaignId:
         "0x6738320fdf80785ff7a1d45ed93a6ffa07068ce9ec4170c1887d09f32fba7b57",
-      version: 2,
     },
   },
 });